file_hash_create and file_hash_check implemented
The decided format for the hashes is:
METHOD:HASHSUM

To be able to test if the created file hashes are correct, a helper
function that evaluates the a file against a supplied hash was
implemented. This calls on the "main" function to check if the hash
generated for that file is the same as the one supplied to the check
function.

The tests just check three of the available hash algorimthms, SHA1,
SHA256, and MD5.