Refact: Archive type is specified as input in main entrypoint
Instead of guessing the archive type to be used from the output
filename, it can be specified directly. All this is done at the
entrypoint cleaning up the logic of the Archiver class.
If no archive type parameter is set as input, the type is still
infered from the filename of the output. Defaults to tar.gz
if neither output name nor archive type is set.