Feat: Separate package-wrapper meta-data into own field.
Since the meta-data from the package-wrapper process is separate
from meta-data put in by the user it should be easy to identify it
in a separate field.

One test is also commented out since the time stamp from creating
is no longer triggered by calling the Manifest class but from
running the entry point.